+++ engine/lite/request-response/runtime/src/main/scala/pl/touk/nussknacker/engine/requestresponse/RequestResponseHttpHandler.scala
@@ -9,7 +9,12 @@ import pl.touk.nussknacker.engine.api.Context
 import pl.touk.nussknacker.engine.api.component.NodeComponentInfo
 import pl.touk.nussknacker.engine.api.exception.NuExceptionInfo
 import pl.touk.nussknacker.engine.lite.api.commonTypes.ErrorType
-import pl.touk.nussknacker.engine.requestresponse.api.{RequestResponseGetSource, RequestResponsePostSource}
+import pl.touk.nussknacker.engine.requestresponse.api.{
+  RequestResponseGetSource,
+  RequestResponsePostSource,
+  ResponseEncoder
+}
+import pl.touk.nussknacker.engine.requestresponse.api.Request
 import pl.touk.nussknacker.engine.util.Implicits.RichScalaMap
 
 import scala.jdk.CollectionConverters._
@@ -25,28 +30,31 @@ class RequestResponseHttpHandler[Effect[_]: Monad](
   // TODO: refactor responseEncoder/source API
   def invoke(request: HttpRequest, entity: Array[Byte]): Effect[Either[NonEmptyList[ErrorType], Json]] = {
     for {
-      input         <- tryInvoke(tryToParse(request, entity))
-      rawResult     <- EitherT(invokeInterpreter(input))
-      encoderResult <- tryInvoke(encoder.toJsonResponse(input, rawResult))
+      parsedRequest <- tryInvoke(tryToParseRequest(request, entity))
+      rawResult     <- EitherT(invokeInterpreter(parsedRequest))
+      encoderResult <- tryInvoke(encoder.toJsonResponse(parsedRequest, rawResult))
     } yield encoderResult
   }.value
 
-  private def tryToParse(request: HttpRequest, entity: Array[Byte]) = {
+  private def tryToParseRequest(request: HttpRequest, entity: Array[Byte]): Request[Any] = {
+    val headers = request.headers.map(h => h.name() -> h.value()).toMap
     (source, request.method) match {
       case (source: RequestResponsePostSource[Any], HttpMethods.POST) =>
-        source.parse(entity)
+        source.parse(entity, headers)
       case (source: RequestResponseGetSource[Any], HttpMethods.GET) =>
         val paramsMultiMap = request.getUri().query().toMultiMap.asScala.toMap.mapValuesNow(_.asScala.toList)
-        source.parse(paramsMultiMap)
+        source.parse(paramsMultiMap, headers)
       case (_, method) =>
         throw new IllegalArgumentException(s"Method $method is not handled")
     }
   }
 
-  private val source  = requestResponseInterpreter.source
-  private val encoder = source.responseEncoder.getOrElse(DefaultResponseEncoder)
+  private val source                        = requestResponseInterpreter.source
+  private val encoder: ResponseEncoder[Any] = source.responseEncoder.getOrElse(DefaultResponseEncoder)
 
-  private def invokeInterpreter(input: Any) = requestResponseInterpreter.invokeToOutput(input).map(_.toEither)
+  private def invokeInterpreter(input: Request[Any]) = {
+    requestResponseInterpreter.invokeToOutput(input).map(_.toEither)
+  }
 
   private def tryInvoke[T](value: => T): EitherT[Effect, NonEmptyList[ErrorType], T] =
     EitherT.fromEither[Effect](