+++ engine/lite/request-response/runtime/src/test/scala/pl/touk/nussknacker/engine/requestresponse/RequestResponseInterpreterSpec.scala
@@ -1,13 +1,13 @@
 package pl.touk.nussknacker.engine.requestresponse
 
-import cats.data.{NonEmptyList, ValidatedNel}
-import cats.data.Validated.{Invalid, Valid}
+import cats.data.Validated.Valid
+import cats.data.ValidatedNel
 import com.typesafe.config.ConfigFactory
 import io.dropwizard.metrics5.MetricRegistry
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import pl.touk.nussknacker.engine.RuntimeMode
-import pl.touk.nussknacker.engine.api.{Context, ContextId, NodeId, ProcessVersion}
+import pl.touk.nussknacker.engine.api.{Context, ContextId, NodeId, ProcessVersion, TraceId}
 import pl.touk.nussknacker.engine.api.component.{ComponentId, ComponentType, NodeComponentInfo, NodesDeploymentData}
 import pl.touk.nussknacker.engine.api.exception.NuExceptionInfo
 import pl.touk.nussknacker.engine.api.runtimecontext.IncContextIdGenerator
@@ -16,10 +16,12 @@ import pl.touk.nussknacker.engine.build.{GraphBuilder, ScenarioBuilder}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.lite.api.commonTypes.ErrorType
 import pl.touk.nussknacker.engine.lite.api.runtimecontext.LiteEngineRuntimeContextPreparer
+import pl.touk.nussknacker.engine.lite.api.utils.sources.BaseLiteSource
 import pl.touk.nussknacker.engine.lite.components.LiteBaseComponentProvider
 import pl.touk.nussknacker.engine.lite.components.requestresponse.RequestResponseComponentProvider
 import pl.touk.nussknacker.engine.lite.metrics.dropwizard.DropwizardMetricsProviderFactory
 import pl.touk.nussknacker.engine.requestresponse.FutureBasedRequestResponseScenarioInterpreter.InterpreterType
+import pl.touk.nussknacker.engine.requestresponse.api.Request
 import pl.touk.nussknacker.engine.resultcollector.ProductionServiceInvocationCollector
 import pl.touk.nussknacker.engine.spel.SpelExtension._
 import pl.touk.nussknacker.engine.testing.LocalModelData
@@ -36,6 +38,8 @@ class RequestResponseInterpreterSpec extends AnyFunSuite with Matchers with Pati
 
   import scala.concurrent.ExecutionContext.Implicits.global
 
+  private val traceId = TraceId.generate().toString
+
   test("run scenario in request response mode") {
 
     val scenario = ScenarioBuilder
@@ -50,7 +54,7 @@ class RequestResponseInterpreterSpec extends AnyFunSuite with Matchers with Pati
     val contextId = firstIdForFirstSource(scenario)
     val result    = runScenario(scenario, Request1("a", "b"), creator)
 
-    result shouldBe Valid(List(Response(s"alamakota-$contextId")))
+    result shouldBe Valid(List(Response(s"alamakota-$contextId-$traceId")))
     creator.processorService.invocationsCount.get() shouldBe 1
   }
 
@@ -87,7 +91,7 @@ class RequestResponseInterpreterSpec extends AnyFunSuite with Matchers with Pati
       val contextId = firstIdForFirstSource(scenario)
       val result    = invokeInterpreter(interpreter, Request1("a", "b"))
 
-      result shouldBe Valid(List(Response(s"alamakota-$contextId")))
+      result shouldBe Valid(List(Response(s"alamakota-$contextId-$traceId")))
       creator.processorService.invocationsCount.get() shouldBe 1
 
       eventually {
@@ -276,7 +280,7 @@ class RequestResponseInterpreterSpec extends AnyFunSuite with Matchers with Pati
       case NuExceptionInfo(
             Some(NodeComponentInfo("sinkId", Some(ComponentId(ComponentType.Sink, "unknown")))),
             SinkException("FailingSink failed"),
-            Context(`contextId`, variables, None),
+            Context(`contextId`, variables, None, Some(traceId)),
             _,
             _
           ) :: Nil if variables == Map("input" -> Request1("a", "b")) =>
@@ -424,8 +428,10 @@ class RequestResponseInterpreterSpec extends AnyFunSuite with Matchers with Pati
     interpreter
   }
 
-  private def invokeInterpreter(interpreter: InterpreterType, input: Any) = {
-    interpreter.invokeToOutput(input).futureValue
+  private def invokeInterpreter(interpreter: InterpreterType, input: Any): ValidatedNel[ErrorType, List[Any]] = {
+    interpreter
+      .invokeToOutput(Request(input, Map(BaseLiteSource.DefaultTraceIdHeader -> traceId)))
+      .futureValue
   }
 
   private def firstIdForFirstSource(scenario: CanonicalProcess): ContextId =